fix: include types/readable-stream as a dependency (#111)
The type definitions for the Metrics client are incomplete without
this dependency, and the fix is hard to identify from the error
messages. This is technically not a runtime dependency, and affects
only a subset of our users, but this makes the module much easier
to consume as TypeScript and/or JSDoc typechecking users.